python: Fix comparison between signed and unsigned integer
Fix these warnings, seen in python 3 builds on Travis:

sanlock.c:1002:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
         if (res->lver == -1) {
                       ^
sanlock.c:1118:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
         if (res->lver == -1) {

+++ python/sanlock.c
@@ -999,7 +999,7 @@ py_acquire(PyObject *self __unused, PyObject *args, PyObject *keywds)
     if (version != Py_None) {
         res->flags |= SANLK_RES_LVER;
         res->lver = pyinteger_as_unsigned_long_long_mask(version);
-        if (res->lver == -1) {
+        if (res->lver == (uint64_t)-1) {
             __set_exception(EINVAL, "Unable to convert the version value");
             goto exit_fail;
         }
@@ -1115,7 +1115,7 @@ py_request(PyObject *self __unused, PyObject *args, PyObject *keywds)
     } else {
         res->flags |= SANLK_RES_LVER;
         res->lver = pyinteger_as_unsigned_long_long_mask(version);
-        if (res->lver == -1) {
+        if (res->lver == (uint64_t)-1) {
             __set_exception(EINVAL, "Unable to convert the version value");
             goto exit_fail;
         }
